Corrected definition of pearsonr for masked arrays
The given formula was incorrect. The t-statistic will only follow a students-t distribution if you ignore each PAIR of values (x_i, y_i) for which at least one of x_i, y_i is masked. This was not done in the denominator, so returning p-values based on that formula is misleading.

Let x' denote the (smaller) array obtained by removing masked values from x, and y' denote y with masked values removed. Further let x'' be x with values removed if they are masked in either x or y, and similarly y''. The old formula, which equated to
    ( len(x'')  / sqrt(len(x') * len(y')) ) * cov(x'', y'') / (var(x') * var(y')),
does not relate to any other definition of correlation.
This version is simply:
   cov(x'', y'') / (std(x'') * std(y'')),
which gives the correctly distributed t-statistic.